requirements.txt
pygame>=2.6.0
pyserial>=3.5
